selection sort algorithm implemented seamlessly! Colors working great too
