+++ src/Algorithms/insertionSort.ts
@@ -1,3 +1,6 @@
+import { swap, createNewTracker, addStepToTracker } from '../helperFunctions';
+import { Tracker } from '../IterationTrackerInterface';
+
 export const insertionSortInfo = {
     name: 'Insertion Sort',
     description: 'Insertion sort is a simple sorting algorithm that iterates through an array and at each iteration it removes one element from the array, finds the location it belongs to in the sorted list and inserts it there, repeating until no elements remain in the unsorted list. It is an in-place, stable sorting algorithm that is inefficient on large input arrays but works well for data sets that are almost sorted. It is more efficient in practice compared to other quadratic sorting algorithms like bubble sort and selection sort.',
