--- src/Algorithms/selectionSort.js
@@ -1,10 +1,0 @@
-export const selectionSortInfo = {
-    name: 'Selection Sort',
-    description: 'Selection sort is an in-place comparison sorting algorithm that divides the input list into two parts: the sublist of items already sorted, which is built up from left to right at the front (left) of the list, and the sublist of items remaining to be sorted that occupy the rest of the list. Initially, the sorted sublist is empty and the unsorted sublist is the entire input list. The algorithm proceeds by finding the smallest element in the unsorted sublist, exchanging (swapping) it with the leftmost unsorted element (putting it in sorted order), and moving the sublist boundaries one element to the right.',
-    performance: {
-        best: 'O(n^2)',
-        average: 'O(n^2)',
-        worst: 'O(n^2)'
-    },
-    index: 1
-}
