+++ src/App.js
@@ -11,8 +11,8 @@ import SortingStepControls from './components/SortingStepControls.js';
 
 // Algorithms
 import {bubbleSortInfo, BubbleSort} from './Algorithms/bubbleSort.ts';
-import {selectionSortInfo} from './Algorithms/selectionSort.js';
-import {insertionSortInfo} from './Algorithms/insertionSort.js';
+import {selectionSortInfo, SelectionSort} from './Algorithms/selectionSort.ts';
+import {insertionSortInfo} from './Algorithms/insertionSort.ts';
 import {quickSortInfo} from './Algorithms/quickSort.js';
 import {mergeSortInfo} from './Algorithms/mergeSort.js';
 import {heapSortInfo} from './Algorithms/heapSort.js';
@@ -35,7 +35,8 @@ const ALGOINFO = [
 ]
 
 const ALGORITHMS = [
-    BubbleSort
+    BubbleSort,
+    SelectionSort
 ]
 
 const ALGONAMES = ALGOINFO.map(algo => algo.name);
@@ -170,7 +171,7 @@ const App = () => {
         console.log(index);
         setAlgorithmInfo(ALGOINFO[index]);
         // setData(generateRandomUniqueUnorderedList(parseInt(dataSize))); // make new state for completed - only do on completed
-        if (ALGOINFO[index].name === 'Bubble Sort' && !algorithmInfo.index) {
+        if (ALGOINFO[index].name === 'Selection Sort' && !algorithmInfo.index) {
             getSpecificAlgorithmTracker(index, sortingData? sortingData.array : data); // for some reason, without this it will send the sorted list - so when you click the same alogirthm twice it solves it automatically. needs investigation but this fixes it
         }
     }
